context menu on main activity
currently this does nothing. need to pass a view or listView onto
registerForContextMenu() method but it only exists within fragments.
keeping it here in case im doing something wrong and there is something
for me to pass in main activity. Also i accidently forget to commit this
in my last commit.